+++ AdapterUnitTests/BidMachineAdapterTests/BidMachineBannerAdTests.swift
@@ -46,23 +46,6 @@ final class BidMachineRTBBannerAdTests {
     }
   }
 
-  @Test("RTB banner ad load fails for missing a bid response")
-  func load_fails_whenBidResponseIsMissing() async {
-    let adConfig = AUTKMediationBannerAdConfiguration()
-    let adapter = BidMachineAdapter()
-
-    await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
-      adapter.loadBanner(for: adConfig) { ad, error in
-        let error = error as NSError?
-        #expect(error != nil)
-        #expect(error!.code == BidMachineAdapterError.ErrorCode.invalidAdConfiguration.rawValue)
-        #expect(ad == nil)
-        continuation.resume()
-        return AUTKMediationBannerAdEventDelegate()
-      }
-    }
-  }
-
   @Test("RTB banner ad load fails for failing to create a request config")
   func load_fails_whenBidMachineFailsToCreateRequestConfig() async {
     client.shouldBidMachineSucceedCreatingRequestConfig = false
@@ -150,6 +133,116 @@ final class BidMachineRTBBannerAdTests {
     let adConfig = AUTKMediationBannerAdConfiguration()
     adConfig.bidResponse = "test response"
     adConfig.watermark = "test watermark".data(using: .utf8)
+    let adapter = BidMachineAdapter()
+    let eventDelegate = AUTKMediationBannerAdEventDelegate()
+    var delegate: BidMachineAdDelegate?
+
+    await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
+      adapter.loadBanner(for: adConfig) { ad, error in
+        let error = error as NSError?
+        #expect(error == nil)
+        #expect(ad != nil)
+        delegate = ad as? BidMachineAdDelegate
+        continuation.resume()
+        return eventDelegate
+      }
+    }
+    delegate?.didTrackInteraction?(
+      OCMockObject.mock(for: BidMachineBanner.self) as! BidMachineBanner)
+
+    #expect(eventDelegate.reportClickInvokeCount == 1)
+  }
+
+}
+
+@Suite("BidMachine adapter waterll banner")
+final class BidMachineWaterfallBannerAdTests {
+
+  let client: FakeBidMachineClient
+
+  init() {
+    client = FakeBidMachineClient()
+    BidMachineClientFactory.debugClient = client
+  }
+
+  @Test("RTB banner ad load succeeds")
+  func load_succeeds() async {
+    let adConfig = AUTKMediationBannerAdConfiguration()
+    let adapter = BidMachineAdapter()
+
+    await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
+      adapter.loadBanner(for: adConfig) { ad, error in
+        let error = error as NSError?
+        #expect(error == nil)
+        #expect(ad != nil)
+        continuation.resume()
+        return AUTKMediationBannerAdEventDelegate()
+      }
+    }
+  }
+
+  @Test("Waterfall banner ad load fails for failing to create a request config")
+  func load_fails_whenBidMachineFailsToCreateRequestConfig() async {
+    client.shouldBidMachineSucceedCreatingRequestConfig = false
+
+    let adConfig = AUTKMediationBannerAdConfiguration()
+    let adapter = BidMachineAdapter()
+
+    await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
+      adapter.loadBanner(for: adConfig) { ad, error in
+        let error = error as NSError?
+        #expect(error != nil)
+        #expect(ad == nil)
+        continuation.resume()
+        return AUTKMediationBannerAdEventDelegate()
+      }
+    }
+  }
+
+  @Test("Waterfall banner ad load fails for failing to create an ad")
+  func load_fails_whenBidMachineFailsToCreateAd() async {
+    client.shouldBidMachineSucceedCreatingAd = false
+
+    let adConfig = AUTKMediationBannerAdConfiguration()
+    let adapter = BidMachineAdapter()
+
+    await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
+      adapter.loadBanner(for: adConfig) { ad, error in
+        let error = error as NSError?
+        #expect(error != nil)
+        #expect(ad == nil)
+        continuation.resume()
+        return AUTKMediationBannerAdEventDelegate()
+      }
+    }
+  }
+
+  @Test("Impression count")
+  func impreesion_count() async {
+    let adConfig = AUTKMediationBannerAdConfiguration()
+    let adapter = BidMachineAdapter()
+    let eventDelegate = AUTKMediationBannerAdEventDelegate()
+    var delegate: BidMachineAdDelegate?
+
+    await withCheckedContinuation { (continuation: CheckedContinuation<Void, Never>) in
+      adapter.loadBanner(for: adConfig) { ad, error in
+        let error = error as NSError?
+        #expect(error == nil)
+        #expect(ad != nil)
+        delegate = ad as? BidMachineAdDelegate
+        continuation.resume()
+        return eventDelegate
+      }
+    }
+    delegate?.didTrackImpression?(
+      OCMockObject.mock(for: BidMachineBanner.self) as! BidMachineBanner)
+
+    #expect(eventDelegate.reportImpressionInvokeCount == 1)
+  }
+
+  @Test("Click count")
+  func click_count() async {
+    let adConfig = AUTKMediationBannerAdConfiguration()
     let adapter = BidMachineAdapter()
     let eventDelegate = AUTKMediationBannerAdEventDelegate()
     var delegate: BidMachineAdDelegate?