+++ adapters/BidMachine/BidMachineAdapter/BannerAdLoader.swift
@@ -53,16 +53,30 @@ final class BannerAdLoader: NSObject, MediationBannerAd, @unchecked Sendable {
   }
 
   func loadAd() {
-    guard let bidResponse = adConfiguration.bidResponse else {
-      handleLoadedAd(
-        nil,
-        error: BidMachineAdapterError(
-          errorCode: .invalidAdConfiguration,
-          description: "The ad configuration is missing bid response."
-        ).toNSError())
-      return
+    if let bidResponse = adConfiguration.bidResponse {
+      loadRTBAd(with: bidResponse)
+    } else {
+      loadWaterfallAd()
     }
 
+  }
+
+  private func loadWaterfallAd() {
+    do {
+      try client.loadWaterfallBannerAd(delegate: self) {
+        [weak self] error in
+        guard let self else { return }
+        guard error == nil else {
+          self.handleLoadedAd(nil, error: error)
+          return
+        }
+      }
+    } catch {
+      handleLoadedAd(nil, error: error)
+    }
+  }
+
+  private func loadRTBAd(with bidResponse: String) {
     guard let watermark = adConfiguration.watermark?.base64EncodedString() else {
       handleLoadedAd(
         nil,
@@ -85,7 +99,6 @@ final class BannerAdLoader: NSObject, MediationBannerAd, @unchecked Sendable {
     } catch {
       handleLoadedAd(nil, error: error)
     }
-
   }
 
   private func handleLoadedAd(_ ad: MediationBannerAd?, error: Error?) {