+++ adapters/BidMachine/CHANGELOG.md
@@ -3,6 +3,7 @@
 #### Next version
 - Add support for waterfall interstitial ads in the BidMachine adapter.
 - Add support for waterfall rewarded ads in the BidMachine adapter.
+- Add support for waterfall banner ads in the BidMachine adapter.
 
 #### [Version 3.4.0.1](https://dl.google.com/googleadmobadssdk/mediation/ios/bidmachine/BidMachineAdapter-3.4.0.1.zip)
 - Test mode is now available in Objective-C through the `GADMediationAdapterBidMachineExtras.isTestMode` property.